% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.he.R
\name{format}
\alias{format}
\title{Format Results}
\usage{
format(result)
}
\arguments{
\item{result}{A list outputted from hie}
}
\description{
This formats the results
}
\examples{
format()
}
\keyword{format}
