% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.he.R
\name{hie}
\alias{hie}
\title{Hierarchical Endpoints}
\usage{
hie(
  endpoints_input,
  sample.size = NA,
  power = NA,
  alpha = 0.05,
  rratio = 0.5,
  output = "ALL"
)
}
\description{
This creates the Probability: win, lose, tie, WR, NB, WO, DOOR and creates sample size or power using results.
}
\examples{
hie()
}
\keyword{endpoints}
