% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.he.R
\name{hie}
\alias{hie}
\title{Hierarchical Endpoints}
\usage{
hie(
  endpoints_input,
  sample.size = NA,
  power = NA,
  alpha = 0.05,
  rratio = 0.5,
  output = "ALL"
)
}
\arguments{
\item{endpoints_input}{A list with each endpoint being a nested list
\itemize{
  \item Time to Event "TTE":
   \itemize{
      \item tte.winning.direction: winning direction ("GT" or "LT")
      \item er.a OR hr: probability of event in group A OR hazard ratio (group A relative to group B)
      \item er.b: probability of event in group B
      \item s: follow-up time
    }
  \item Continuous "Continuous":
   \itemize{
      \item continuous.winning.direction: winning direction ("GT" or "LT")
      \item mu.a OR mean.diff: mean in group A OR mean difference of group A minus group B
      \item mu.b: mean in group B
      \item sd.a: standard deviation in group A
      \item sd.b: standard deviation in group B
      \item delta: threshold to win
      }
  \item Binary (1/0) "Binary":
   \itemize{
      \item binary.winning.direction: winning direction ("GT" or "LT")
      \item pi.a OR prob.diff: Prob(Y=1) in group A OR Prob(Y=1) of group A minus group B
      \item pi.b: Prob(Y=1) in group B
      }
  \item Count Endpoint (such as # of events) "Count":
   \itemize{
      \item count.winning.direction: winning direction ("GT" or "LT")
      \item lam.a OR rr: number of counts/events in group A OR relative rate of group A over group B
      \item lam.b: number of counts/events in group B
      }
  \item Ordinal (1, 2, ..., J) "Ordinal":
   \itemize{
      \item ordinal.winning.direction: winning direction ("GT" or "LT")
      \item pi.ordinal.a: Prob(Y=1), ..., Prob(Y=J) in group A (comma-separated)
      \item pi.ordinal.b: Prob(Y=1), ..., Prob(Y=J) in group B (comma-separated)
      }
}}

\item{sample.size}{An integer (enter either sample.size or power)}

\item{power}{0 to 1 (enter either sample.size or power)}

\item{alpha}{Two-sided Type 1 Error}

\item{rratio}{Randomization probability for Group A}

\item{output}{Choose from: ALL, WR, WO, NB, DOOR}
}
\description{
This creates the probability: win, lose, tie, WR, NB, WO, DOOR and creates sample size or power using results.
}
\examples{
# Two continuous hierarchical endpoints:
# The marginal distributions for Y1A and Y1B are normal distributions with
# means 15 and 4, respectively, and standard deviations of 60. For Y2A and
# Y2B, the marginal distributions are normal distributions with means 40 and
# 30, respectively, and standard deviations of 24. For both endpoints, the
# threshold to win is chosen to be the same, with both delta1 and delta2
# equal to 5.

endpoints_input <- list(
  list(type = "Continuous",
       mu.a = 15,
       mu.b = 4,
       sd.a = 60,
       sd.b = 60,
       delta = 5,
       continuous.winning.direction = "GT"),
  list(type = "Continuous",
       mu.a = 40,
       mu.b = 30,
       sd.a = 24,
       sd.b = 24,
       delta = 5,
       continuous.winning.direction = "GT")
)
hie(endpoints_input,
    power = 0.85,
    alpha = 0.05,
    rratio = 0.5,
    output = "ALL")

# Two binary hierarchical endpoints:
# The marginal probabilities for Y1A and Y1B are binomial distributions with
# a success probability of 0.90 and 0.85, respectively, for one trial. For
# Y2A and Y2B, the marginal probabilities are binomial distributions with
# success probabilities of 0.80 and 0.75, respectively, for one trial.

endpoints_input <- list(
  list(type = "Binary",
      pi.a = 0.9,
      pi.b = 0.85,
      binary.winning.direction = "GT"),
  list(type = "Binary",
      pi.a = 0.8,
      pi.b = 0.75,
      binary.winning.direction = "GT")
)
hie(endpoints_input,
    power = 0.85,
    alpha = 0.05,
    rratio = 0.5,
    output = "ALL")

# Binary and continuous hierarchical endpoints:
# The marginal probabilities for Y1A and Y1B are binomial distributions with
# success probabilities of 0.96 and 0.95, respectively, for one trial. For
# Y2A and Y2B, the marginal distributions are normal distributions with means
# 36 and 31, respectively, and standard deviations of 24.

endpoints_input <- list(
  list(type = "Binary",
       pi.a = 0.96,
       pi.b = 0.95,
       binary.winning.direction = "GT"),
  list(type = "Continuous",
       mu.a = 36,
       mu.b = 31,
       sd.a = 24,
       sd.b = 24,
       delta = 5,
       continuous.winning.direction = "GT")
)
hie(endpoints_input,
    power = 0.85,
    alpha = 0.05,
    rratio = 0.5,
    output = "ALL")

# Time to death and number of hospitalizations as hierarchical endpoints:
# The marginal distributions for Y1A and Y1B are exponential distributions
# with rate parameters of 0.16 and 0.20, respectively. For Y2A, the marginal
# distribution is a Poisson distribution with a mean of 0.75, and for Y2B, it
# is a normal distribution with a mean of 1.1. The follow-up time for all
# measurements is 5 years.

endpoints_input <- list(
  list(type = "TTE",
       tte.winning.direction = "GT",
       s = 5,
       hr.a = 0.8,
       er.b = 0.63212),
  list(type = "Count",
       count.winning.direction = "LT",
       lam.a = 0.75,
       lam.b = 1.1)
)
hie(endpoints_input,
    power = 0.85,
    alpha = 0.05,
    rratio = 0.5,
    output = "ALL")

# Two ordinal hierarchical endpoints, each with 3 ordinal categories:
# The marginal distributions for Y1A and Y1B are multinomial distributions
# with probabilities for the three categories (1, 2, 3) given by
# (0.45, 0.30, 0.25) for Y1A and (0.50, 0.30, 0.20) for Y1B. For Y2A and Y2B,
# the marginal distributions are multinomial distributions with probabilities
# (0.30, 0.30, 0.40) for Y2A and (0.40, 0.30, 0.30) for Y2B. The probabilities
# represent the likelihood of a subject being in categories 1, 2, or 3. We
# assume that a subject in a higher ordinal category wins over a subject in a
# lower ordinal category.

endpoints_input <- list(
  list(type = "Ordinal",
       pi.ordinal.a = c(0.45, 0.3, 0.25),
       pi.ordinal.b = c(0.5, 0.3, 0.2),
       ordinal.winning.direction = "GT"),
  list(type = "Ordinal",
       pi.ordinal.a = c(0.3, 0.3, 0.4),
       pi.ordinal.b = c(0.4, 0.3, 0.3),
       ordinal.winning.direction = "GT")
)
hie(endpoints_input,
    power = 0.85,
    alpha = 0.05,
    rratio = 0.5,
    output = "ALL")
}
\keyword{endpoints}
